+++ src/main/java/proguard/classfile/io/kotlin/KotlinMetadataWriter.java
@@ -450,6 +450,13 @@ implements ClassVisitor,
                 );
             }
 
+            if (kotlinPropertyMetadata.syntheticMethodForDelegate != null)
+            {
+                ext.visitSyntheticMethodForDelegate(
+                    toKotlinJvmMethodSignature(kotlinPropertyMetadata.syntheticMethodForDelegate)
+                );
+            }
+
             ext.visitEnd();
 
             kmdPropertyVisitor.visitEnd();